Added 2D model creation capability, removed show edges toggle button, set name of model to title of photo gallery activity.

+++ trunk/trkr/trunk/ImagesTo3DModels/app/src/main/java/ece473/trekker/imagesto3dmodels/TriangleVertex.java
@@ -34,7 +34,7 @@ public class TriangleVertex extends Point3
      * @return - a new TriangleVertex
      */
     public static TriangleVertex buildTriangleVertex( int face, int row, int column, int maxRow,
-                                                      int maxColumn, int depth )
+                                                      int maxColumn, double depth )
     {
         switch( face )
         {
